Add files via upload
Since we have not gone over what he is expecting for the final paper, this is a heavily outlined version of what it could look like. I did not type in most of the new information because I don't want to have to change it if he tells us he's looking for something else. If you got Q's or suggestions, just let me know